[integrated signup and login api in frontend]
--> first when user signup using the signup form created in frontend it'll take username, password and created
jwt token in backend and return that token and i have stored that token in the localstorage and session storage of the browser,
and then create user object with all its details api will call which will take the jwt token in their header using interceptor that i have created then only that api will work and
user will be created in backend database.

--> when user login then the login api will take the username and password and check in the database that if that username and password
is their in the database, basically it'll authenticate the user and then return the jwt token back in the response so that they can access other apis.